refactor(api): move feedback serializer
